pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.133.1",
    "langchain>=1.2.10",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=1.0.9",
    "livekit-agents>=1.4.3",
    "livekit-api>=1.1.0",
    "openai>=2.24.0",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
